# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API LIVE TEST",
        "description": "Test API V1",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/product/list": {
            "get": {
                "tags": [
                    "site_product"
                ],
                "summary": "site_product_list",
                "operationId": "site_product_list",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 0
                    },
                    {
                        "name": "take",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuthSite": []
                    }
                ]
            }
        },
        "/api/v1/profile": {
            "get": {
                "tags": [
                    "site_profile"
                ],
                "summary": "profile",
                "operationId": "profile",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Returns when customer is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuthSite": []
                    }
                ]
            }
        },
        "/api/v1/profile/update": {
            "put": {
                "tags": [
                    "site_profile"
                ],
                "summary": "profile_update",
                "operationId": "profile_update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "customer@gmail.com"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "name"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "family"
                                    },
                                    "fax": {
                                        "type": "string",
                                        "example": "25215425"
                                    },
                                    "gender": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "16272802467397.png"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "25215425"
                                    },
                                    "mobile": {
                                        "type": "string",
                                        "example": "058454147"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Returns when customer is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuthSite": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Admin",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "Passport",
                "scheme": "bearer"
            },
            "bearerAuthSite": {
                "type": "http",
                "description": "Customer",
                "name": "AuthorizationSite",
                "in": "header",
                "bearerFormat": "Passport",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}